+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 <a name="Pending Release"></a>
-## [Pending Release](https://github.com/lightstep/lightstep-tracer-java-common/compare/master...0.14.1)
+## [Pending Release](https://github.com/lightstep/lightstep-tracer-java-common/compare/master...0.15.0)
+
+<a name="0.15.0"></a>
+## [0.15.0](https://github.com/lightstep/lightstep-com.lightstep.tracer-java-common/compare/0.15.0...0.14.3)
+
+* Upgraded to io.opentracing 0.31.0
+
+### BREAKING CHANGES
+* BaseSpan and ActiveSpan are simplified into a single Span class.
+* Scope replaces ActiveSpan, removing the continuation concept.
+* ScopeManager replaces ActiveSpanSource
+* ThreadLocalScopeManager replaces ThreadLocalActiveSpanSource
 
 <a name="0.14.1"></a>
 ## [0.14.1](https://github.com/lightstep/lightstep-tracer-java-common/compare/0.14.0...0.14.1)